move routes around to new routes files, update all the settings/config, add test api route
