+++ Pipfile
@@ -25,15 +25,15 @@ mccabe = "==0.6.1"
 monotonic = "==1.3"
 oauth2client = "==4.1.2"
 ply = "==3.8"
-protobuf = "==3.4.0"
-pyasn1 = "==0.3.7"
-pyasn1-modules = "==0.1.5"
+protobuf = "==3.6.0"
+pyasn1 = "==0.4.1 "
+pyasn1-modules = "==0.2.1"
 pycodestyle = "==2.3.1"
 pyflakes = "==1.6.0"
 pymongo = "==3.5.1"
 requests = "==2.20.0"
 rsa = "==3.4.2"
-six = "==1.11.0"
+six = "==1.13.0"
 tenacity = "==4.5.0"
 urllib3 = "==1.24.2"
 webassets = "==0.12.1"
@@ -45,6 +45,8 @@ Jinja2 = "==2.10.1"
 MarkupSafe = "==1.0"
 pyScss = "==1.3.5"
 Werkzeug = "==0.15.3"
+python-dotenv = "*"
+google-cloud-bigquery = "*"
 
 [requires]
 python_version = "3.7"
