+++ app.yaml
@@ -17,19 +17,9 @@
 # for details.
 
 runtime: python37
-entrypoint: gunicorn -b :$PORT main:app
+instance_class: F1
 
-#[START cloudsql_settings]
-beta_settings:
-    # If using Cloud SQL, uncomment and set this value to the Cloud SQL
-    # connection name, e.g.
-    #   "project:region:cloudsql-instance"
-    # You must also update the values in config.py.
-    #
-    # cloud_sql_instances: "your-cloudsql-connection-name"
-#[END cloudsql_settings]
 
-# [START handlers]
 handlers:
 - url: /static
   static_dir: aqandu/static
